+++ modules/nano_eval/nano_eval_bridge.h
@@ -1,0 +1,6 @@
+#ifndef NANO_EVAL_BRIDGE_H
+#define NANO_EVAL_BRIDGE_H
+
+#include "nano_eval.h"
+
+#endif